Don't accept up-to-eof as raw_string_content, it breaks recovery
The idea was an unmatched R"delim( should consume the rest of the file.
However there's a side-effect: on *any* parse error tree-sitter likes to
recover by matching the rest of the file as raw_string_content, even
with no delimeter at all.

In this example, everything after `c` becomes raw_string_content:
  a b c d e
  void foo();